Send back new CSRF token after login/logout
Previously, the CSRF token worked, but Devise deletes the token after
you log in or out. So you'd have to refresh the page before making
further POST requests.

Now, we send back the new CSRF token in the XHR response to login or
logout. We intercept that in axios to set the client-side CSRF token to
the new value.

Got info from:

* https://recalll.co/app/?q=ajax%20-%20Rails,%20Devise%20authentication,%20CSRF%20issue
* http://stackoverflow.com/questions/11845500/rails-devise-authentication-csrf-issue