Cache thin documents for super-fast header rendering!
The index page loads a thin version of all documents and puts them in
redux as a cache (it never uses this cache itself, that introduces a bit
too much complexity and isn't very useful).

The document show page uses the cache as a first version of its
`document` data. This means that it can show the title and editable
state of the document in the header, before the document has fully
loaded.

This might be nice if you have large documents with a ton of data.