Use rack-cors globally for OPTIONS, GET
Should fix CORS issues for Rails 5.

Downside: applies globally (not just for API endpoints like I want).
But, low risk of problems since this is just for GET requests.

Upside: works.

Will add to the planning board and revisit later to see if I can scope
to just the API.